backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0
asyncpg==0.30.0
aiosqlite==0.20.0
httpx==0.28.1
anthropic==0.43.0
apscheduler==3.10.4
python-dotenv==1.0.1
pydantic==2.10.4
pydantic[email]==2.10.4
pydantic-settings==2.7.1
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
yfinance>=0.2.36
pykrx>=1.0.45
cryptography>=43.0.0
dart-fss>=0.4.10
neo4j>=5.27.0
pytest>=8.0
pytest-asyncio>=0.23
holidays>=0.60  # KR 공휴일 회피 (배당 ex-date T-1 영업일 계산)
